+++ src/index.js
@@ -30,6 +30,7 @@ import matchPath from './utils/route-utils.js';
 import AuditsController from './controllers/audits.js';
 import OrganizationsController from './controllers/organizations.js';
 import SitesController from './controllers/sites.js';
+import HooksController from './controllers/hooks.js';
 import SlackController from './controllers/slack.js';
 import trigger from './controllers/trigger.js';
 
@@ -81,6 +82,7 @@ async function run(request, context) {
   try {
     const routeHandlers = getRouteHandlers(
       AuditsController(context.dataAccess),
+      HooksController(context),
       OrganizationsController(context.dataAccess, log),
       SitesController(context.dataAccess, log),
       SlackController(SlackApp),